Allow docker-compose down to "fail" as the services might have significantly changes
When testing against branch `294-load-data-after-up`

If you run

```
docker-compose -f eclipse-pass.base.yml -f eclipse-pass.demo.yml down
```

If will fail because the compose changed _soooo much_.

```
ERROR: The Compose file is invalid because:
Service ember has neither an image nor a build context specified. At least one must be provided.
```

So it should be OK for the "down" to fail and not fail the deploy.